refactor(pipeline): decouple topology rules into subsystem handlers
