+++ src/ramses_rf/eavesdropper.py
@@ -11,8 +11,19 @@ from datetime import timedelta as td
 from typing import TYPE_CHECKING, Any
 
 import ramses_rf.exceptions as exc
-from ramses_rf.const import SZ_TEMPERATURE, SZ_ZONE_IDX
+from ramses_rf.const import (
+    I_,
+    RQ,
+    SZ_DOMAIN_ID,
+    SZ_TEMPERATURE,
+    SZ_ZONE_IDX,
+    ZON_ROLE_MAP,
+    DevType,
+)
 from ramses_rf.devices import Device, Temperature
+from ramses_rf.enums import TopologyAction
+from ramses_rf.models import TopologyChangedEvent
+from ramses_rf.protocol.ramses import CODES_ONLY_FROM_CTL, HVAC_KLASS_BY_VC_PAIR
 from ramses_tx import Code
 from ramses_tx.address import HGI_DEV_ADDR
 from ramses_tx.typing import DeviceIdT
@@ -33,6 +44,22 @@ class EavesdropEngine:
         """
         self._gwy: Gateway = gwy
         self._prev_30c9_map: dict[str, Message] = {}
+
+    def _emit(self, event: TopologyChangedEvent) -> None:
+        """Emit a heuristic topology event to the DeviceRegistry."""
+        registry = getattr(self._gwy, "device_registry", None)
+        if registry:
+            registry.handle_topology_event(event)
+
+    def _get_payloads(self, msg: Message) -> list[Any]:
+        """Safely extract the array or standard dictionary payload."""
+        raw: Any = getattr(msg, "payload", getattr(msg, "data", {}))
+        if isinstance(raw, dict):
+            res = raw.get("_array", [raw])
+            return res if isinstance(res, list) else [res]
+        if isinstance(raw, list):
+            return raw
+        return []
 
     def eavesdrop_referenced_devices(self, msg: Message) -> None:
         """Instantiate implicitly referenced header devices during eavesdropping.
@@ -77,13 +104,7 @@ class EavesdropEngine:
                     registry.get_device(addr_id)
 
     def _get_tcs(self, msg: Message) -> Any:
-        """Find the active TCS read-model instance.
-
-        :param msg: The incoming message.
-        :type msg: Message
-        :returns: The active TCS instance if found.
-        :rtype: Any
-        """
+        """Find the active TCS read-model instance."""
         if tcs := getattr(msg.src, "tcs", None):
             return tcs
         if tcs := getattr(msg.dst, "tcs", None):
@@ -124,8 +145,17 @@ class EavesdropEngine:
         tcs = self._get_tcs(msg)
 
         # 1. Telemetry-driven Sensor/Actuator Linking
-        # NOTE: Do NOT prematurely bind msg.src.id to zone based on telemetry,
-        # as 000C binding packets and temperature correlation handle authoritative parenting.
+        self._evaluate_evohome_rules(msg)
+        self._evaluate_directed_telemetry_rules(msg)
+        self._evaluate_ufh_communication_rules(msg)
+        self._evaluate_hvac_rules(msg)
+        self._evaluate_dhw_opentherm_rules(msg)
+        self._evaluate_heating_prefix_rules(msg)
+        self._evaluate_appliance_control_sync_rules(msg)
+        self._evaluate_appliance_eavesdrop_rules(msg)
+        self._evaluate_zone_type_eavesdrop_rules(msg)
+        self._evaluate_implicit_binding_rules(msg)
+        self._evaluate_third_address_broadcast_rules(msg)
 
         # 2. 30C9 Temperature Broadcast Correlation
         if msg.code == Code._30C9 and tcs:
@@ -138,37 +168,375 @@ class EavesdropEngine:
             elif getattr(msg.src, "type", None) in ("03", "04", "12", "22", "34"):
                 await self._eavesdrop_from_trv_broadcast(tcs, msg)
 
+    def _evaluate_evohome_rules(self, msg: Message) -> None:
+        """Evaluate rules specific to the Evohome CH/DHW ecosystem."""
+        if msg.verb == I_ and msg.code in CODES_ONLY_FROM_CTL:
+            self._emit(
+                TopologyChangedEvent(
+                    action=TopologyAction.CREATE_CONTROLLER,
+                    device_id=msg.src.id,
+                    causation="Rule_Evohome_Controller_Broadcast",
+                )
+            )
+
+    def _evaluate_directed_telemetry_rules(self, msg: Message) -> None:
+        """Evaluate implicit bindings from directed telemetry broadcasts."""
+        ctl_id = None
+        if getattr(msg.dst, "type", None) == "01":
+            ctl_id = msg.dst.id
+        elif getattr(msg.addr3, "type", None) == "01":
+            ctl_id = msg.addr3.id
+
+        if msg.verb == I_ and ctl_id and msg.src.id != ctl_id:
+            for p in self._get_payloads(msg):
+                if not isinstance(p, dict):
+                    continue
+
+                zone_idx = p.get(SZ_ZONE_IDX)
+                domain_id = p.get(SZ_DOMAIN_ID)
+
+                if zone_idx is None and domain_id is None:
+                    continue
+
+                metadata: dict[str, Any] = {}
+
+                is_actuator = getattr(msg.src, "type", None) in ("04", "08", "13", "02")
+                is_sensor = getattr(msg.src, "type", None) in (
+                    "00",
+                    "03",
+                    "07",
+                    "12",
+                    "22",
+                    "34",
+                )
+
+                if msg.code in (Code._3150, Code._0008, Code._2309, Code._000A):
+                    metadata["device_role"] = "actuator"
+                elif msg.code in (
+                    Code._30C9,
+                    Code._1260,
+                    Code._10A0,
+                    Code._12B0,
+                ):
+                    metadata["device_role"] = "sensor" if is_sensor else "actuator"
+                    if is_sensor:
+                        metadata["is_sensor"] = "True"
+                else:
+                    metadata["device_role"] = "actuator" if is_actuator else "sensor"
+                    if is_sensor:
+                        metadata["is_sensor"] = "True"
+
+                if zone_idx is not None:
+                    metadata["zone_idx"] = str(zone_idx)
+                    metadata["child_id"] = str(zone_idx)
+                elif domain_id is not None:
+                    if domain_id in ("F9", "FA", "FC"):
+                        metadata["domain_id"] = str(domain_id)
+                        metadata["child_id"] = str(domain_id)
+                    else:
+                        metadata["zone_idx"] = str(domain_id)
+                        metadata["domain_id"] = str(domain_id)
+                        metadata["child_id"] = str(domain_id)
+
+                self._emit(
+                    TopologyChangedEvent(
+                        action=TopologyAction.BIND_DEVICE,
+                        parent_id=ctl_id,
+                        child_id=msg.src.id,
+                        metadata=metadata,
+                        causation=f"Rule_Telemetry_Eavesdrop_{msg.code}",
+                    )
+                )
+
+    def _evaluate_ufh_communication_rules(self, msg: Message) -> None:
+        """Evaluate rules specific to Underfloor Heating (UFH) communication eavesdropping."""
+        is_ufc_src = getattr(msg.src, "type", None) == "02"
+        is_ufc_dst = getattr(msg.dst, "type", None) == "02"
+
+        if not (is_ufc_src or is_ufc_dst):
+            return
+
+        ufc_id = msg.src.id if is_ufc_src else msg.dst.id
+
+        ctl_id = None
+        if getattr(msg.src, "type", None) == "01":
+            ctl_id = msg.src.id
+        elif getattr(msg.dst, "type", None) == "01":
+            ctl_id = msg.dst.id
+        elif getattr(msg.addr3, "type", None) == "01":
+            ctl_id = msg.addr3.id
+
+        if ctl_id and ctl_id != ufc_id:
+            event_promote = TopologyChangedEvent(
+                action=TopologyAction.UPDATE_DEVICE_CLASS,
+                device_id=ufc_id,
+                metadata={"device_class": DevType.UFC},
+                causation="Rule_UFH_Communication_Promotion",
+            )
+            self._emit(event_promote)
+
+            event_bind = TopologyChangedEvent(
+                action=TopologyAction.BIND_DEVICE,
+                parent_id=ctl_id,
+                child_id=ufc_id,
+                metadata={"device_role": "ufc"},
+                causation="Rule_UFH_Communication_Binding",
+            )
+            self._emit(event_bind)
+
+    def _evaluate_hvac_rules(self, msg: Message) -> None:
+        """Evaluate rules specific to Ventilation and HVAC eavesdropping."""
+        msg_verb = msg.verb
+        msg_code = str(msg.code)
+
+        dev_class = None
+        for (schema_verb, schema_code), dev_class_name in HVAC_KLASS_BY_VC_PAIR.items():
+            if (schema_verb is None or schema_verb == msg_verb) and str(
+                schema_code
+            ) == msg_code:
+                dev_class = dev_class_name
+                break
+
+        if dev_class:
+            if msg.src.id != "--:------" and getattr(msg.src, "type", None) != "01":
+                self._emit(
+                    TopologyChangedEvent(
+                        action=TopologyAction.UPDATE_DEVICE_CLASS,
+                        device_id=msg.src.id,
+                        metadata={"device_class": dev_class},
+                        causation="Rule_HVAC_Signature_Source",
+                    )
+                )
+
+            if (
+                msg.dst.id != "--:------"
+                and msg.dst.id != msg.src.id
+                and getattr(msg.dst, "type", None) != "01"
+            ):
+                self._emit(
+                    TopologyChangedEvent(
+                        action=TopologyAction.UPDATE_DEVICE_CLASS,
+                        device_id=msg.dst.id,
+                        metadata={"device_class": dev_class},
+                        causation="Rule_HVAC_Signature_Target",
+                    )
+                )
+
+    def _evaluate_dhw_opentherm_rules(self, msg: Message) -> None:
+        """Evaluate rules specific to DHW and OpenTherm Bridges."""
+        if msg.code == Code._3220 and getattr(msg.src, "type", None) == "10":
+            self._emit(
+                TopologyChangedEvent(
+                    action=TopologyAction.UPDATE_DEVICE_CLASS,
+                    device_id=msg.src.id,
+                    metadata={"device_class": DevType.OTB},
+                    causation="Rule_OTB_3220_Signature",
+                )
+            )
+
+        elif (
+            msg.code in (Code._1260, Code._10A0)
+            and getattr(msg.src, "type", None) == "07"
+        ):
+            self._emit(
+                TopologyChangedEvent(
+                    action=TopologyAction.UPDATE_DEVICE_CLASS,
+                    device_id=msg.src.id,
+                    metadata={"device_class": DevType.DHW},
+                    causation="Rule_DHW_Signature",
+                )
+            )
+
+    def _evaluate_heating_prefix_rules(self, msg: Message) -> None:
+        """Evaluate passive heuristics based purely on hardware prefixes."""
+        prefix_map = {
+            "03": DevType.HCW,
+            "04": DevType.TRV,
+            "12": DevType.THM,
+            "13": DevType.BDR,
+            "22": DevType.THM,
+            "34": DevType.THM,
+        }
+
+        addrs = [msg.src]
+        if msg.dst.id != "--:------" and msg.dst.id != msg.src.id:
+            addrs.append(msg.dst)
+
+        for addr in addrs:
+            if getattr(addr, "type", None) in prefix_map:
+                self._emit(
+                    TopologyChangedEvent(
+                        action=TopologyAction.UPDATE_DEVICE_CLASS,
+                        device_id=addr.id,
+                        metadata={"device_class": prefix_map[addr.type]},
+                        causation="Rule_Heating_Prefix_Heuristic",
+                    )
+                )
+
+    def _evaluate_appliance_control_sync_rules(self, msg: Message) -> None:
+        """Evaluate direct configuration syncs to map System Relay."""
+        if (
+            getattr(msg.src, "type", None) == "01"
+            and msg.dst.id != "--:------"
+            and getattr(msg.dst, "type", None) == "13"
+        ):
+            if msg.code in (Code._1100, Code._10E0, Code._1FC9):
+                self._emit(
+                    TopologyChangedEvent(
+                        action=TopologyAction.BIND_DEVICE,
+                        parent_id=msg.src.id,
+                        child_id=msg.dst.id,
+                        metadata={
+                            "domain_id": "FC",
+                            "child_id": "FC",
+                            "device_role": "appliance_control",
+                        },
+                        causation="Rule_Direct_Relay_Sync",
+                    )
+                )
+
+    def _evaluate_appliance_eavesdrop_rules(self, msg: Message) -> None:
+        """Evaluate legacy passive eavesdropping of System Relay."""
+        if msg.code not in (Code._3220, Code._3B00, Code._3EF0):
+            return
+
+        app_cntrl_id: DeviceIdT | None = None
+
+        if msg.code == Code._3220 and msg.verb == RQ:
+            if (
+                getattr(msg.src, "type", None) == "01"
+                and msg.dst.id != "--:------"
+                and getattr(msg.dst, "type", None) == "10"
+            ):
+                app_cntrl_id = msg.dst.id
+
+        elif msg.code == Code._3EF0 and msg.verb == RQ:
+            if (
+                getattr(msg.src, "type", None) == "01"
+                and msg.dst.id != "--:------"
+                and getattr(msg.dst, "type", None) in ("10", "13")
+            ):
+                app_cntrl_id = msg.dst.id
+
+        if app_cntrl_id is not None:
+            self._emit(
+                TopologyChangedEvent(
+                    action=TopologyAction.BIND_DEVICE,
+                    parent_id=msg.src.id,
+                    child_id=app_cntrl_id,
+                    metadata={
+                        "domain_id": "FC",
+                        "child_id": "FC",
+                        "device_role": "appliance_control",
+                    },
+                    causation="Rule_Legacy_Appliance_Eavesdrop",
+                )
+            )
+
+    def _evaluate_zone_type_eavesdrop_rules(self, msg: Message) -> None:
+        """Evaluate legacy passive promotion of zone classes."""
+        for p in self._get_payloads(msg):
+            if not isinstance(p, dict):
+                continue
+
+            zone_idx = p.get(SZ_ZONE_IDX)
+            if zone_idx is None:
+                continue
+
+            zone_class: str | None = None
+
+            if msg.code in (Code._0008, Code._0009):
+                zone_class = ZON_ROLE_MAP["ELE"]
+
+            if zone_class is not None:
+                ctl_id = None
+                if getattr(msg.dst, "type", None) == "01":
+                    ctl_id = msg.dst.id
+                elif getattr(msg.src, "type", None) == "01":
+                    ctl_id = msg.src.id
+                elif getattr(msg.addr3, "type", None) == "01":
+                    ctl_id = msg.addr3.id
+
+                if ctl_id is not None:
+                    self._emit(
+                        TopologyChangedEvent(
+                            action=TopologyAction.UPDATE_TRAITS,
+                            device_id=ctl_id,
+                            metadata={
+                                "zone_idx": str(zone_idx),
+                                "class": zone_class,
+                            },
+                            causation="Rule_Legacy_Zone_Type_Eavesdrop",
+                        )
+                    )
+
+    def _evaluate_implicit_binding_rules(self, msg: Message) -> None:
+        """Evaluate implicit bindings from directed controller polls."""
+        if msg.verb not in (RQ, "W"):
+            return
+
+        if getattr(msg.src, "type", None) != "01":
+            return
+
+        if msg.dst.id == "--:------":
+            return
+
+        dst_type = getattr(msg.dst, "type", None)
+        if dst_type not in ("00", "04", "08"):
+            return
+
+        self._emit(
+            TopologyChangedEvent(
+                action=TopologyAction.BIND_DEVICE,
+                parent_id=msg.src.id,
+                child_id=msg.dst.id,
+                metadata={
+                    "device_role": "actuator" if dst_type in ("04", "08") else "sensor"
+                },
+                causation="Rule_Implicit_Poll_Binding",
+            )
+        )
+
+    def _evaluate_third_address_broadcast_rules(self, msg: Message) -> None:
+        """Evaluate bindings from 3rd address field of broadcasts."""
+        if msg.verb != I_:
+            return
+
+        src_type = getattr(msg.src, "type", None)
+        if getattr(msg.addr3, "type", None) == "01" and src_type in ("00", "04", "08"):
+            self._emit(
+                TopologyChangedEvent(
+                    action=TopologyAction.BIND_DEVICE,
+                    parent_id=msg.addr3.id,
+                    child_id=msg.src.id,
+                    metadata={
+                        "device_role": "actuator"
+                        if src_type in ("04", "08")
+                        else "sensor"
+                    },
+                    causation="Rule_3rd_Address_Declaration",
+                )
+            )
+
     async def _eavesdrop_from_controller_broadcast(
         self, tcs: Any, msg: Message, prev: Message | None
     ) -> None:
-        """Correlate recently heard TRV temperatures against a new zone array.
-
-        :param tcs: The target TCS read-model.
-        :type tcs: Any
-        :param msg: The incoming 30C9 array message from the controller.
-        :type msg: Message
-        :param prev: The previous 30C9 array message to check changes.
-        :type prev: Message | None
-        """
+        """Correlate recently heard TRV temperatures against a new zone array."""
         if prev is None:
             return
 
-        # TODO: use msgz/I, not RP
         secs = await tcs.entity_state.get_value(Code._1F09, key="remaining_seconds")
         if not isinstance(secs, (int, float)):
             secs = 300.0
         if msg.dtm > prev.dtm + td(seconds=secs + 5):
-            # can only compare against 30C9 pkt from the last cycle
             return
 
-        # zones with changed temps
         changed_zones: dict[str, float] = {
             z.get(SZ_ZONE_IDX): z.get(SZ_TEMPERATURE)
             for z in msg.payload
             if z not in prev.payload and z.get(SZ_TEMPERATURE) is not None
         }
         if not changed_zones:
-            # ctl's 30C9 says no zones have changed temps during this cycle
             return
 
         def _testable_zones(chg_zones: dict[str, float]) -> dict[float, str]:
@@ -187,7 +555,7 @@ class EavesdropEngine:
 
         testable_zones = _testable_zones(changed_zones)
         if not testable_zones:
-            return  # no testable zones
+            return
 
         testable_sensors_map: dict[float, list[Device]] = {}
         for d in self._gwy.device_registry.devices:
@@ -198,7 +566,6 @@ class EavesdropEngine:
                         testable_sensors_map[d_temp] = []
                     testable_sensors_map[d_temp].append(d)
 
-        # COLLISION ABSTENTION: Drop temperatures reported by multiple sensors unless one is dedicated
         unique_sensors: dict[float, Device] = {}
         for temp_val, devs in testable_sensors_map.items():
             if len(devs) == 1:
@@ -211,7 +578,7 @@ class EavesdropEngine:
                     unique_sensors[temp_val] = dedicated[0]
 
         if not unique_sensors:
-            return  # no unique testable sensors available, must abstain
+            return
 
         matched_pairs = {
             sensor: zone_idx
@@ -235,17 +602,15 @@ class EavesdropEngine:
                     sensor.id, parent=zone, is_sensor=True
                 )
 
-        # now see if we can allocate the controller as a sensor...
         if any(z for z in tcs.zones if z.sensor is tcs.ctl):
-            return  # the controller is already a sensor
+            return
 
         remaining_zones = _testable_zones(changed_zones)
         if len(remaining_zones) != 1:
-            return  # no testable zones
+            return
 
         temp, zone_idx = tuple(remaining_zones.items())[0]
 
-        # can safely(?) assume this zone is using the CTL as a sensor...
         if not [s for s in unique_sensors if s == temp]:
             zone = tcs.zone_by_idx[zone_idx]
             with contextlib.suppress(
@@ -258,13 +623,7 @@ class EavesdropEngine:
                 )
 
     async def _eavesdrop_from_trv_broadcast(self, tcs: Any, msg: Message) -> None:
-        """Correlate a new TRV temperature broadcast against known zones.
-
-        :param tcs: The target TCS read-model.
-        :type tcs: Any
-        :param msg: The incoming 30C9 message from a TRV.
-        :type msg: Message
-        """
+        """Correlate a new TRV temperature broadcast against known zones."""
         if not isinstance(msg.payload, dict):
             return
 
@@ -283,7 +642,6 @@ class EavesdropEngine:
                 if zone_temp == trv_temp:
                     matching_zones.append(zone)
 
-        # COLLISION ABSTENTION: Bind only if exactly one zone matches this temp
         if len(matching_zones) == 1:
             with contextlib.suppress(
                 exc.DeviceNotFoundError,
